+++ NEWS.md
@@ -6,7 +6,7 @@
 * export new `string_repr()` generic that is used to layout R objects for 
   formatted log message output (#48, thanks @mmuurr)
   
-* The `$log()` method of Logger and LoggerGlue now unpacks conditons (except
+* The `$log()` method of Logger and LoggerGlue now unpacks conditions (except
   if they are supplied as a named argument) (#45, thanks @mmuurr)
   
 * Fix some timezone related tests for CRAN
@@ -74,7 +74,7 @@
 * `Logger$log()` dispatches to all appenders - even if some throw an error -
   instead of aborting after the first Appender that throws an error
   
-* complete rewrite of the documentation to use the new roxgen2 features for
+* complete rewrite of the documentation to use the new roxygen2 features for
   R6 classes.