+++ NEWS.md
@@ -2,9 +2,13 @@
 
 * `%k` and `%K` parameters in `format.LogEvent` now work as expected when using
   **crayon** terminal colours (#52).
+  
 * Fix default format string for `LayoutGlue` which was using `msg` 
   instead of `{msg}` (#54, thx @mmuurr)  
   
+* Update docs to use the more common term "structured logging" instead of 
+  "custom fields" where appropriate
+
 
 # lgr 0.4.3